+++ CHANGELOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 - Backup AES Key File with database backup. (#243)
 - Copy AES Key File on database copy. (#244)
 - Delete AES Key File on database delete. (#245)
+- Restore AES Key File on database restoration. (#246)
 
 ### Bug Fixes